Commiting 8th task of springboot

+++ src/main/java/com/stackroute/serivce/MuzixServiceImpl.java
@@ -1,15 +1,51 @@
 package com.stackroute.serivce;
 
 import com.stackroute.domain.Muzix;
+import com.stackroute.exceptions.TrackAlreadyExistsException;
+import com.stackroute.exceptions.TrackNotFoundException;
 import com.stackroute.repository.MuzixRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.context.ApplicationListener;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.event.ContextRefreshedEvent;
 import org.springframework.stereotype.Service;
+import sun.tools.java.Environment;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
-public class MuzixServiceImpl implements MuzixService {
+@PropertySource("application.properties")
+public class MuzixServiceImpl implements MuzixService, ApplicationListener<ContextRefreshedEvent>, CommandLineRunner {
+    @Autowired
+    Environment environment;
+    //Adding the default values to the database by using Application listener
+    @Value("${muzix.1.id:default}")
+    int id1;
+    @Value("${muzix.1.name:default}")
+    String name1;
+    @Value("${muzix.1.track:default}")
+    String track1;
+    @Value("${muzix.2.id:default}")
+    int id2;
+    @Value("${muzix.2.name:default}")
+    String name2;
+    @Value("${muzix.2.track:default}")
+    String track2;
+    //Overriding the commandlinerunner method
+    @Override
+    public void run(String... args) throws Exception {
+        System.out.println("command line running before application starts");
+    }
+    //Overriding application context method
+    @Override
+    public void onApplicationEvent(ContextRefreshedEvent contextRefreshedEvent) {
+        System.out.println("Running ContextRefreshedEvent");
+        muzixRepository.save(new Muzix(id1,name1,track1));
+        muzixRepository.save(new Muzix(id2, name2, track2));
+    }
     //Autowiring the muzix repository
     @Autowired
     MuzixRepository muzixRepository;
@@ -21,8 +57,12 @@ public class MuzixServiceImpl implements MuzixService {
 
     //creating the track
     @Override
-    public Muzix saveMuzix(Muzix muzix) {
+    public Muzix saveMuzix(Muzix muzix) throws TrackAlreadyExistsException {
 
+        if(muzixRepository.existsById(muzix.getId()))
+        {
+            throw new TrackAlreadyExistsException("Track already exists");
+        }
         return muzixRepository.save(muzix);
     }
 
@@ -34,11 +74,14 @@ public class MuzixServiceImpl implements MuzixService {
 
     //Deleting the track
     @Override
-    public boolean deleteMuzix(int id) {
+    public boolean deleteMuzix(int id) throws TrackNotFoundException {
+        if(!muzixRepository.existsById(id))
+        {
+            throw new TrackNotFoundException("Track not found");
+        }
         muzixRepository.deleteById(id);
         return true;
     }
-
     //Updating the existing track
     @Override
     public Muzix updateMuzix(Muzix muzix) {
